backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-filter==25.1
python-dotenv==1.1.1
Pillow==11.3.0
celery[redis]==5.5.3
redis==5.0.4
